V3.1.5 custom (#1)
* Custom code to generate charts and graphs.

Added the code for the SD graph and findings chart. Added custom functions to Jinja env.

* Added custom modules to install process.

Forgot to add the custom modules to the requirements text file.

* Fixed install with new dependencies.

Updated Dockerfiles and downgraded scipy because of a bug with recent version.

* Fixed threading issue.

Fixed code after testing in GW when using custom code. Importing the template works with custom code but currently not rendering as images in report.

* Chart and graph integration.

Fixed the code to integrate the findings chart and SD graph correctly in the template. Changed font family for generated images to use open source "Arial Narrow" equivalent. Added jinja sort_findings filter to sort findings correctly in report.

* Added custom enums file. Updated code to handle richtext for our custom fields.

Enum to house our custom code to minimize as much as we can touching core code. Updated fields we are using for retest status and DoE to use RichText.

* Fixed comparison.

Fixed the lookup.

* Fixed some bugs.

Got the styling working for the Difficulty Exploit field.

* Code to dynamically calculate SD score.

Integrated the formula to calculate the SD score and plot in graph.

* Finding Chart Dynamic data integrated.

Added the code to add dynamic data for the finding chart. Currently running into an issue with overlapping labels.

* Got the findings chart fully working.

Fixed x-axis overlapping issue, blurry pictures in report, and fixed a bug in the chart generation where HTML came from GhostWriter field.

* Adjusted to make SD graph bigger.

Made some adjustments to the height to make the SD graph bigger after some feedback.

* Added pie chart for netsec and modified report pages.

Added code for pie chart generation and modified the labels to match what John currently has in GW.

* Updated Z-Score on retests.

Z-Score (SD score) was updated for retests to handle the status of the finding.

* Custom report findings page.

Added the ability to search report findings based on criteria to learn from others and gauge performance.

* Added pagination and menu item for custom page.

* Added the finding type icon to page.

Added the finding type icon to page to match what we currently have in GW.

* Fixed formatting of files.

---------

Co-authored-by: Colin McQueen <colin.mcqueen@stratumsecurity.com>