V4.3.11 custom tmp (#23)
* V4.3.1 custom code (#22)

* Merge custom code from v4.2.5 into latest version (#21)

Fixed conflicts.

* Added back our customizations to the page.

* Fixed table on page.

Fixed the header on the page.

* Fixed a bug when the grade is an F.

The original if statement wasn't handling F grades correctly. Think it had to do with 0 value when the grade is an F.

* Fixed name conflict.

GW added a Severity class that conflicted with our enum. Had to create an alias for our enum to fix the conflict.

* Renamed the file to be accurate and fixed reports with no findings.

The new code that supported cloud finding types broke the reports with 0 findings. Had to rely on grabbing all the finding types to calculate the averages for the report template to grab the grade.